Patch spec to report mapped ports correctly
The spec is wrong, and supporting the right thing required extending
the handlers a little (this will be needed to support 1.32 anyway).
I've added a new sample response, which also has issues.  At least
this time the response is directly out of the server (rather than
suffering a yaml -> R -> json translation).