Sample responses for 1.34
This uses the json files from the docker website, because pulling these
out of the spec looks challenging now that they're spread out across
all the parts of the response.  On the plus side, this makes the responses
easier to read and means that we can avoid any odd vector/scalar conversion
issues when doing the translations.

The other downside is that they're quite large!  We might need to move some
of this into a release artefact and download on demand rather than
including in the actual package or it might get a bit big